chore: require integer values

+++ src/modules/beers/beers.entity.ts
@@ -23,8 +23,8 @@ export const BasicSearchEntity = z.object({
 });
 
 export const AdvancedSearchEntity = z.object({
-  page: z.coerce.number().min(1).optional(),
-  per_page: z.coerce.number().min(1).max(80).optional(),
+  page: z.coerce.number().int().min(1).optional(),
+  per_page: z.coerce.number().int().min(1).max(80).optional(),
   abv_gt: z.coerce.number().optional(),
   abv_lt: z.coerce.number().optional(),
   ibu_gt: z.coerce.number().optional(),
